token_server/requirements.txt
fastapi
python-jose
uvicorn[standard]
